Rework the X implementation to use only one X connection
